Stacked membranes and solvate per lipid
This update introduces one new subcommand for solvations, which allows the number of solvent that should be placed within a solvent box to be calculated from the number of lipids present within the solvent box. See the documentation under "solvation" for how to use it.

The update also introduces a new "special" command called "stacked_membranes" which allows one to systematically create stacked membranes, with it being easy to control the number of stacked membranes, their composition, the distance between the individual membranes, and the composition of the solvent between each set of membranes. See the new section "Special commands" in the documentation pdf and the new tutorial system entry titled "Multiple Bilayers System V2" for explanations of how to use it.